requirements.txt
discord.py
pandas
apscheduler